+++ src/adf_json_processor/config/config.py
@@ -1,116 +1,170 @@
+import json
 import os
 import datetime
-import ast
+from utils.logger import Logger
 
-class Config:
+class ConfigManager:
     """
-    Handles configuration and paths for processing ADF data.
+    Manages configuration for Azure Data Factory (ADF) pipelines by setting up Databricks widgets
+    and managing directory paths for logging and outputs. Provides utility functions to handle
+    directory structure and configuration validation.
     """
-    def __init__(self, auth_strategy, debug=False):
-        self.debug = debug
 
-        if 'dbutils' not in globals() or dbutils is None:
-            raise ValueError("dbutils is required to retrieve secrets or widgets in this environment.")
-
-        # Parse ADF configuration
-        adf_config_str = dbutils.widgets.get("ADFConfig")
-        adf_config = ast.literal_eval(adf_config_str)
-        self.organization, self.project, self.repository, self.branch, self.folder_path = adf_config
-
-        self.source_storage_account = dbutils.widgets.get("SourceStorageAccount")
-        self.destination_storage_account = dbutils.widgets.get("DestinationStorageAccount")
-        self.datasetidentifier = dbutils.widgets.get("Datasetidentifier")
-        self.source_filename = dbutils.widgets.get("SourceFileName")
-
-        self.auth_strategy = auth_strategy
-
-        self.adf_details = {
-            "Organization": self.organization,
-            "Project": self.project,
-            "Repository": self.repository,
-            "Branch": self.branch,
-            "Folder Path": self.folder_path
-        }
-
-        self.log_path = self.generate_log_path()
-        self.output_path = self.generate_output_path()
-
-    def update_source_filename(self, source_filename):
+    def __init__(self, dbutils, logger=None):
         """
-        Ensure the source filename ends with .json.
+        Initializes the ConfigManager with dbutils for widget access and an optional logger.
+        
+        Args:
+            dbutils: Databricks utility object for accessing widgets.
+            logger (Logger): Optional logger instance for structured logging.
         """
-        if not source_filename.endswith(".json"):
-            source_filename += ".json"
-        self.source_filename = source_filename
+        if not dbutils:
+            raise ValueError("dbutils is required to manage widgets in this environment.")
+        
+        self.dbutils = dbutils
+        self.config = {}
+        self.logger = logger or Logger()
 
-    def generate_log_path(self):
+    def initialize_widgets(self):
         """
-        Generate log path based on the current timestamp.
+        Sets up Databricks widgets for ADF pipeline configuration and other storage-related parameters.
+        """
+        self.logger.log_start("initialize_widgets")
+        self.dbutils.widgets.text("adfConfig", '["energinet", "DataPlatform_v3.0", "data-factory", "main", "pipeline"]', "ADF Configuration")
+        self.dbutils.widgets.text("sourceStorageAccount", "dplandingstoragetest", "Source Storage Account")
+        self.dbutils.widgets.text("destinationStorageAccount", "dpuniformstoragetest", "Destination Storage Account")
+        self.dbutils.widgets.text("datasetIdentifier", "xazms", "Dataset Identifier")
+        self.dbutils.widgets.text("sourceFileName", "*", "Source File Name")
+        self.logger.log_end("initialize_widgets")
+
+    def load_config(self):
+        """
+        Loads and validates widget values, stores them in the config attribute, logs the configuration settings,
+        ensures directories exist, and logs optional configuration paths.
+        """
+        self.logger.log_start("load_config")
+        try:
+            # Load and validate configuration
+            self._load_main_config()
+            self.logger.log_block("Loaded Configuration", [f"{k}: {v}" for k, v in self.config.items()])
+
+            # Log optional paths right after main configuration
+            self._log_optional_paths()
+
+            # Ensure directories exist
+            self.ensure_directories_exist()
+        except (ValueError, json.JSONDecodeError) as e:
+            self.logger.log_error(f"Configuration loading error: {e}")
+            raise
+        except Exception as e:
+            self.logger.log_error(f"Unexpected error during configuration loading: {e}")
+            raise ValueError("Unexpected error during configuration loading.")
+        self.logger.log_end("load_config")
+
+    def _load_main_config(self):
+        """
+        Retrieves and parses main configuration values from widgets and stores them in `self.config`.
+        """
+        adf_config_str = self.dbutils.widgets.get("adfConfig")
+        self.config["adfConfig"] = self._parse_and_validate_adf_config(adf_config_str)
+        self.config["sourceStorageAccount"] = self.dbutils.widgets.get("sourceStorageAccount")
+        self.config["destinationStorageAccount"] = self.dbutils.widgets.get("destinationStorageAccount")
+        self.config["datasetIdentifier"] = self.dbutils.widgets.get("datasetIdentifier")
+        self.config["sourceFileName"] = self.dbutils.widgets.get("sourceFileName")
+
+    def _parse_and_validate_adf_config(self, adf_config_str):
+        """
+        Parses and validates the adfConfig JSON string.
+        
+        Args:
+            adf_config_str (str): JSON string containing ADF configuration.
+        
+        Returns:
+            list: Parsed configuration as a list if valid.
+        
+        Raises:
+            ValueError: If the configuration is invalid.
+        """
+        adf_config = json.loads(adf_config_str)
+        if len(adf_config) < 5:
+            self.logger.log_error("ADF Configuration requires 5 elements.")
+            raise ValueError("ADF Configuration requires 5 elements: organization, project, repository, branch, and folder path.")
+        return adf_config
+
+    def _generate_log_path(self):
+        """
+        Generates a log path based on the current timestamp.
+        
+        Returns:
+            str: Path to the log file.
         """
         date_str = datetime.datetime.now().strftime("%Y%m%d_%H%M%S")
-        return f"/dbfs/mnt/{self.source_storage_account}/{self.datasetidentifier}/log/error_log_{date_str}.json"
+        return f"/dbfs/mnt/{self.config['sourceStorageAccount']}/{self.config['datasetIdentifier']}/log/error_log_{date_str}.json"
 
-    def generate_output_path(self):
+    def _generate_output_path(self):
         """
-        Generate output path for the processed file.
+        Generates an output path for the processed file.
+        
+        Returns:
+            str: Path to the output file.
         """
-        return f"/dbfs/mnt/{self.source_storage_account}/{self.datasetidentifier}/combined_hierarchical_pipeline_structure_filtered.json"
-
-    def print_params(self):
-        """
-        Print configuration parameters if debug is enabled.
-        """
-        if not self.debug:
-            return
-
-        print("\n=== ADF Configuration ===")
-        for key, value in self.adf_details.items():
-            print(f"{key}: {value}")
-
-        print("\n=== Storage Configuration ===")
-        print(f"Source Storage Account: {self.source_storage_account}")
-        print(f"Destination Storage Account: {self.destination_storage_account}")
-        print(f"Dataset Identifier: {self.datasetidentifier}")
-        print(f"Source Filename: {self.source_filename}")
-
-        print("\n=== Authentication ===")
-        print(f"Authentication Method: {self.auth_strategy.__class__.__name__}")
-        if isinstance(self.auth_strategy, PATAuthStrategy):
-            print(f"Personal Access Token: {'*' * len(self.auth_strategy.pat)}")
+        return f"/dbfs/mnt/{self.config['sourceStorageAccount']}/{self.config['datasetIdentifier']}/combined_hierarchical_pipeline_structure_filtered.json"
 
     def ensure_directories_exist(self):
         """
-        Ensure the required directories exist in the file system.
+        Ensures that required directories exist in the file system.
+        Logs whether each directory already exists or was created.
         """
-        log_dir = os.path.dirname(self.log_path)
-        output_dir = os.path.dirname(self.output_path)
-        
-        if 'dbutils' in globals():
-            try:
-                dbutils.fs.mkdirs(log_dir)
-                dbutils.fs.mkdirs(output_dir)
-            except Exception as e:
-                print(f"An error occurred while creating directories in DBFS: {e}")
-        else:
-            try:
-                if not os.path.exists(log_dir):
-                    os.makedirs(log_dir)
-                if not os.path.exists(output_dir):
-                    os.makedirs(output_dir)
-            except Exception as e:
-                print(f"An error occurred while creating directories locally: {e}")
+        self.logger.log_start("ensure_directories_exist")
+        log_dir, output_dir = os.path.dirname(self._generate_log_path()), os.path.dirname(self._generate_output_path())
+        dir_statuses = self._check_and_create_directories(log_dir, output_dir)
+        self.logger.log_block("Directory Statuses", dir_statuses)
+        self.logger.log_end("ensure_directories_exist")
 
-def initialize_config(auth_strategy, debug=False):
-    """
-    Initialize configuration with authentication and debug flag.
-    
-    Args:
-        auth_strategy (object): Authentication strategy (PAT or OAuth).
-        debug (bool): Flag for enabling debug mode.
-    
-    Returns:
-        Config: Config object with initialized settings.
-    """
-    config = Config(auth_strategy=auth_strategy, debug=debug)
-    config.print_params()  # Print configuration parameters if in debug mode
-    return config
+    def _check_and_create_directories(self, log_dir, output_dir):
+        """
+        Helper method to check and create directories as needed.
+        
+        Args:
+            log_dir (str): Path for the log directory.
+            output_dir (str): Path for the output directory.
+        
+        Returns:
+            list: Status messages indicating whether directories existed or were created.
+        """
+        dir_statuses = []
+        if 'dbutils' in globals():  # Databricks environment
+            for dir_path, dir_type in [(log_dir, "Log"), (output_dir, "Output")]:
+                try:
+                    self.dbutils.fs.ls(dir_path)
+                    dir_statuses.append(f"{dir_type} directory already exists in DBFS: {dir_path}")
+                except Exception:
+                    self.dbutils.fs.mkdirs(dir_path)
+                    dir_statuses.append(f"{dir_type} directory created in DBFS: {dir_path}")
+        else:  # Local environment
+            for dir_path, dir_type in [(log_dir, "Log"), (output_dir, "Output")]:
+                if os.path.exists(dir_path):
+                    dir_statuses.append(f"{dir_type} directory already exists locally: {dir_path}")
+                else:
+                    os.makedirs(dir_path)
+                    dir_statuses.append(f"{dir_type} directory created locally: {dir_path}")
+        return dir_statuses
+
+    def _log_optional_paths(self):
+        """
+        Logs additional configuration paths for optional outputs.
+        """
+        optional_config = {
+            "log_path": self._generate_log_path(),
+            "output_path": self._generate_output_path()
+        }
+        self.logger.log_block("Optional Configuration", [f"{k}: {v}" for k, v in optional_config.items()])
+
+    def get_config(self):
+        """
+        Returns the configuration dictionary.
+        
+        Returns:
+            dict: Configuration dictionary.
+        """
+        return self.config
